Move app chat template tests into unit suite
